unit tests for disk usage
